services/transaction_engine/requirements.txt
Django==4.2.6
djangorestframework==3.14.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pika==1.3.2
django-cors-headers==4.3.1
django-redis==5.4.0
django-enumchoicefield==3.0.1
requests==2.31.0 